Fixed chunk transfer to save memory.
Now chunk transfer is processed one by one to save memory use for
send/recv buffer. There is a little concern on the latency issue,
which may not be a problem as long as the chunk size is large
enough (> MB?).
Also fixed to resize particle transfer buffer.
Yet the code sometimes crashes as the current implementation uses
a lot more memory than is actually needed.